Add a `set_to_zero` transformation.
The `set_to_zero` transformation maps input gradients to zero. As mentioned in
issue #175, which proposed this feature, this transformation is useful for
example in combination with `multi_transform` as it allows keeping certain
model parameters fixed when the updates are applied.

This PR also adds a note to the docstring of the identity transform to avoid
confusion with `set_to_zero`. The former leaves gradients unchanged, the latter
maps gradients to zero so that the parameters are unchanged when the
gradient-based updates are applied.

PiperOrigin-RevId: 406153776