Fix the eps_root in AdaBelief
This now matches the official PyTorch implementation and the latest revision of the paper on arxiv (the current implementation matched the equations in earlier versions of the paper where the epsilon was not added to the second moment of the prediction error).

Also updated default values of hyper-params.

This subsumes a previous internal PR (https://github.com/deepmind/optax/pull/171/files).

PiperOrigin-RevId: 407819213