Test that lookahead passes parameters correctly.
Fixes issue [#184](https://github.com/deepmind/optax/issues/184).

The simple test optimizer in `lookahead_test` did not use the parameters and
therefore did not catch a bug in the way the parameters were passed to the fast
optimizer [#165](https://github.com/deepmind/optax/pull/165).
This commit changes the test to catch these kinds of bugs in the future.

PiperOrigin-RevId: 408289307