Do not reuse mini-batches between epochs in DPSGD example.
As pointed out in #149, the dataset in the DPSGD example was converted to a
list which meant that the same mini-batches were used for each epoch. This is
fixed now so that the data is shuffled differently in each epoch.

The change led to a small (20%) performance decrease when running on a GPU but this
could be solved by prefetching batches.

PiperOrigin-RevId: 411027866