Fix EMA bias correction
EMA bias correction was erroneously being applied to the EMA state. If debiased EMA gets applied to a constant sequence, the output should remain constant, but because the EMA state was getting "debiased" at every iteration, the output would instead explode and then shrink. This fix applies the bias correction only to the return values and not to the state, matching the formulation from https://arxiv.org/abs/1412.6980.

PiperOrigin-RevId: 425376411