Start inject_hyperparams step count at 0.
This fixes issue #415. Previously, `inject_hyperparams` started with
`step_count=1` in the first update when using schedules (it incremented before
passing it to the schedule) whereas `scale_by_schedule` started with
`step_count=0`. To make this consistent, this PR changes `inject_hyperparams`
to also start at 0, i.e. increment the count only after passing it to the
schedule.

The PR comes with a test that breaks without the change. Furthermore, the step
counts in the existing tests of `inject_hyperparams` had to be decremented by
one in order for the tests to pass.

PiperOrigin-RevId: 479553657