Avoid one layer of indirection by colocating the source and the subpackage __init__ for optax.contrib.
WARNING: This can potentially break the loading of a saved optimizer states if the path of the MechanicState class is tied to the saved data (e.g. via pickling). This kind of pickling issue can be fixed at load time by adding an import stub at the old location, or by adding a `sys.modules` entry for the old path.
PiperOrigin-RevId: 570397989