Use sphinx-contributors for an automated contributors list.
I think having an automated list of contributors is a good idea to give
credit to everyone who has contributed to the project without wasting
time compiling this list manually.

It's debatable whether commit count is the best metric to use for this
but IMO it's the best metric we have that's easily accessible.