Created a NTXent loss
This is the NTXent loss from the SimClr paper. I have done some minor testing on it and it reproduces both loss and gradient values of pytorch_metric_learning's function NTXentLoss. Let me know if you want me to add a test function as well. I was unsure if I should put it in classification because it is really a self-supervised loss, so let me know if you want it somewhere else!