LBFGS solver part 2: implementing linesearch ensuring sufficient decrease and small curvature
It is recommended to combine LBFGS with a linesearch procedure that ensures not only sufficient decrease but also small curvature along the chosen update direction. We implement the zoom_linesearch presented by Nocedal and Wright, used as reference linesearch method in scipy.

The user API follows the one of the scale_by_backtracking_linesearch, i.e., (i) it is a GradientTransformExtraArgs that requires access to value, grad and value_fn, (ii) it can be combined with any base GradientTransform, (iii) value and gradient computed during the linesearch can be recycled using the optax.value_and_grad_from_state utility.

PiperOrigin-RevId: 643033634