Ensure it's possible to donate both the schedule-free state and the params
Prior to this commit, with a schedule-free optimizer, if you tried to donate
an entire train state, including both the parameters and a freshly initialized
optimizer state, you'd get an "INVALID_ARGUMENT: Attempt to donate the same
buffer twice in Execute()" error, because z was the same array as params. This
commit fixes the issue and adds a test.